Revert moving splitter to hardware
