+++ nengo_loihi/splitter.py
